Update main.cpp: Set options at runtime
the smti executable now uses the following command line options:

    $ smti <size> <pref_length> <tie_density> <name>

  <size> is the number of agents on either side
  <pref_length> is the length of preferences on one side
  <tie_density> is the tie density on either side
  <name> is the prefix to use when saving the instance as a string
  (<name>.instance) or a WCNF formatted file (<name>.wcnf)

Note that the lengths of preferences on the second set of agents is
determined by the preferences of the first set of agents. That is, the
second set of agents considers every agent compatible, and simply
randomly shuffles the agents, removes any incompatible pairs, and then
adds ties as per the tie density.