Add model with merged constraints
Adds a model with merged stability constraints. Merging constraints
reduces the number of constraints, but instead increases the number of
variables in a model. The trade-off is better if there are few ranks (or
many ties) in an instance, but the exact point at which this trade off
is efficient is not known.